Updated CSS for image container and download link.
Made changes to the CSS for the image container and download link to improve the visual appearance and functionality of the page. Specifically, I added padding and background color to the image container, as well as hover effects for the child elements. I also positioned and styled the download link to be more prominent on hover, and added a blur filter effect to the image when the download link is hovered over.