Added single image download functionality with the download icon.
This commit adds the ability to download single images from the image gallery with a download icon.
The code uses the Font Awesome library to display the icon and the 'download' attribute in the anchor tag to enable image downloading.
The CSS styling is also included in the 'style.css' file.