This commit initializes the Mapbox map by creating a new instance of the map object. The access token is provided to authenticate the Mapbox API. The map is set to display the streets-v11 style and centered on India's coordinates with an initial zoom level of 5. Navigation controls are added to the map to enable zooming and panning.
An event listener is also added to the map to capture the click event. When a user clicks on the map, the longitude and latitude coordinates of the selected city are logged to the console. Further actions can be performed with the selected city, such as adding a marker or storing the coordinates for future use.